Clean up hotkeys + move lock button to actions
