+++ src/app/item-actions/LockButton.m.scss.d.ts
